+++ backend/Application/Api/GraphQL/Subscription.cs
@@ -7,5 +7,10 @@ namespace Application.Api.GraphQL;
 public class Subscription
 {
     [Subscribe]
+    public AccountsUpdatedSubscriptionItem AccountsUpdated(
+        [Topic] string accountAddress,
+        [EventMessage] AccountsUpdatedSubscriptionItem message) => message;
+
+    [Subscribe]
     public Block BlockAdded([EventMessage] Block block) => block;
 }